backend/requirements.txt
# API Framework
fastapi==0.128.7
uvicorn[standard]==0.34.2
pydantic==2.11.3
pydantic-settings==2.9.1

# Database
sqlalchemy[asyncio]==2.0.46
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1

# Data Processing
pandas==3.0.0
numpy==2.2.4

# Technical Analysis
TA-Lib==0.6.8

# ML
xgboost==3.2.0
scikit-learn==1.7.0
joblib==1.5.0

# AWS
boto3==1.42.45

# HTTP & WebSocket
httpx==0.28.1
websockets==15.0.1
aiohttp==3.12.6

# Utilities
python-dotenv==1.0.1
